buffer: rever to using a 32-bit bufsize_t

+++ src/buffer.h
@@ -13,7 +13,7 @@
 extern "C" {
 #endif
 
-typedef ssize_t bufsize_t;
+typedef int32_t bufsize_t;
 
 typedef struct {
   cmark_mem *mem;
